pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2024-2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
# SPDX-License-Identifier: Apache-2.0
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64"]

[project]
name = "vss_ctx_rag"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
license-files = ["LICENSE", "LICENSE.3rdparty"]
dependencies = [
  "openai>=1.68.2,<2.0.0",
  "langchain_core>=0.3.72",
  "langchain>=0.3.27",
  "langchain_community>=0.3.27",
  "langchain_milvus>=0.2.1",
  "langchain-openai>=0.3.10",
  "langchain-experimental>=0.3.4",
  "langchain-nvidia-ai-endpoints>=0.3.13",
  "pymilvus>=2.5.14",
  "pymilvus-model==0.3.2",
  "python-multipart==0.0.18",
  "pydantic>=2.11.7",
  "pyyaml>=6.0.2",
  "protobuf>=5.29.5",
  "redis==5.2.1",
  "unstructured[all-docs]>=0.15.1",
  "uvicorn[standard]==0.32.0",
  "fastapi>=0.115.4,<0.116.0",
  "requests>=2.32.3,<2.33.0",
  "jsonschema>=4.22.0,<4.23.0",
  "schema>=0.7.7,<0.8.0",
  "json-repair==0.30.2",
  "opentelemetry-sdk>=1.28.2",
  "opentelemetry-api>=1.28.2",
  "opentelemetry-exporter-otlp-proto-http>=1.28.2",
  "opentelemetry-instrumentation-fastapi>=0.49b2",
  "nvtx==0.2.11",
  "matplotlib==3.9.4",
  "torch>=2.0.0,<3.0.0",
  "safetensors==0.4.4",
  "minio==7.2.15",
  "pyaml-env>=1.2.1,<2.0.0",
  "bleach==6.2.0",
  "dataclass-wizard==0.27.0",
  "pdfplumber>=0.6",
  "nvidia-rag>=2.2.2",
  "openinference-semantic-conventions>=0.1.21",
  "openinference-instrumentation-openai>=0.1.30",
  "openinference-instrumentation-langchain>=0.1.44",
  "neo4j>=5.24",
  "fastmcp>=2.11.3",
  "langgraph>=0.6.4",
  "langchain-elasticsearch>=0.3.2"
]

[project.urls]
Homepage = "https://github.com/NVIDIA/context-aware-rag"
Documentation = "https://nvidia.github.io/context-aware-rag/index.html"
Repository = "https://github.com/NVIDIA/context-aware-rag"

[tool.setuptools.packages.find]
where = ["src"]
include = ["vss_ctx_rag.*"]

[tool.setuptools.package-data]
"vss_ctx_rag" = ["data/*"]

[tool.uv.workspace]
members = ["packages/*"]

[tool.setuptools]
include-package-data = true


[tool.uv]
managed = true
config-settings = { editable_mode = "compat" }
package = true

[tool.uv.sources]
vss_ctx_rag_arango = { workspace = true }
vss_ctx_rag_nat = { workspace = true }

[project.optional-dependencies]
docs = [
  "ipython~=8.31",
  "myst-parser~=4.0",
  "nbsphinx~=0.9",
  "nvidia-sphinx-theme>=0.0.7",
  "sphinx~=8.2",
  "sphinx-copybutton>=0.5",
  "sphinx-autoapi>=3.6",
  "vale==3.9.5",
  "setuptools-scm>=8.1.0",
  "sphinxcontrib-mermaid>=1.0.0"
]

arango = ["vss_ctx_rag_arango"]
nat = ["vss_ctx_rag_nat"]
